refactor(vue): align the Vue target with main's conventions
- Rename @dunky.dev/state-machine-vue -> @dunky.dev/vue-state-machine,
  matching the <substrate>-state-machine naming every other target uses.
- Version 0.3.2, in step with the fixed @dunky.dev/* group.
- Ship src alongside dist (parity with #53).
- Drop the ComponentEffects alias — react removed it; every target
  exports the ComponentEffect tuple alone.
- Kebab-case the sandbox SFC filenames and format, fixing the lint and
  format CI failures.
- Add vue to the substrate enumerations in AGENTS/ARCHITECTURE.
