+++ AGENTS.md
@@ -5,7 +5,7 @@ repo. This file is the canonical entry point: read it first, every time.
 
 This is Dunky's state-machine monorepo: UI behavior authored once as
 plain TypeScript state machines (`packages/core`), rendered anywhere
-through thin per-substrate targets (`react`, `native`, `opentui`), with
+through thin per-substrate targets (`react`, `vue`, `native`, `opentui`), with
 a benchmark suite, per-substrate sandboxes, and the docs website
 alongside.
 
@@ -24,12 +24,12 @@ editing files in that scope — it overrides anything here for that scope
 
 ## Scopes
 
-| Scope     | Path          | What it is                                                                         |
-| --------- | ------------- | ---------------------------------------------------------------------------------- |
-| Packages  | `packages/**` | The core machine, substrate targets (react, native, opentui), and shared internals |
-| Benchmark | `benchmark/`  | Perf suite comparing against competitor libraries                                  |
-| Sandbox   | `sandbox/`    | Per-substrate demo apps for manual verification                                    |
-| Website   | `website/`    | The docs site                                                                      |
+| Scope     | Path          | What it is                                                                              |
+| --------- | ------------- | --------------------------------------------------------------------------------------- |
+| Packages  | `packages/**` | The core machine, substrate targets (react, vue, native, opentui), and shared internals |
+| Benchmark | `benchmark/`  | Perf suite comparing against competitor libraries                                       |
+| Sandbox   | `sandbox/`    | Per-substrate demo apps for manual verification                                         |
+| Website   | `website/`    | The docs site                                                                           |
 
 Some changes are cross-scope: a change in `core/` may need follow-up in
 the targets, sandboxes, and docs — and vice versa. Check what else your
