+++ ARCHITECTURE.md
@@ -29,7 +29,7 @@ The host
                                |  bridged per target
                                v
 +------------------------------------------------------------------------+
-|  <target>   (react, native, opentui, …)                                |
+|  <target>   (react, vue, native, opentui, …)                                |
 |  Runtime-specific bridge                                               |
 |  • lifecycle (build + start/stop)  • normalize bindings -> props       |
 |  • selector subscription                                               |
@@ -58,7 +58,7 @@ actions. Nothing in `core/` knows that React or the DOM exists.
 substrate-agnostic event and attr vocabulary (`onPress`, `role`, …); `shared/utils`
 owns cross-target helpers (mergeProps, composeHandlers, positioning).
 
-**`<target>/`** is the substrate side — `react`, `native`, `opentui`, and any
+**`<target>/`** is the substrate side — `react`, `vue`, `native`, `opentui`, and any
 future renderer. Each target is the runtime bridge for one environment: the
 lifecycle bridge, the event normalization, and the selector subscription all
 live here.
@@ -93,12 +93,12 @@ Zag, whose machines read props directly.)
 
 ## Project structure
 
-| File / location             | What it owns                                                  |
-| --------------------------- | ------------------------------------------------------------- |
-| `packages/core/`            | State-machine engine (plain-mutation kernel)                  |
-| `packages/shared/bindings/` | Substrate-agnostic event + attr vocabulary (onPress, role, …) |
-| `packages/shared/utils/`    | mergeProps, composeHandlers, positioning, memo                |
-| `packages/<target>/`        | Hook + normalize per substrate (react, native, opentui, …)    |
+| File / location             | What it owns                                                    |
+| --------------------------- | --------------------------------------------------------------- |
+| `packages/core/`            | State-machine engine (plain-mutation kernel)                    |
+| `packages/shared/bindings/` | Substrate-agnostic event + attr vocabulary (onPress, role, …)   |
+| `packages/shared/utils/`    | mergeProps, composeHandlers, positioning, memo                  |
+| `packages/<target>/`        | Hook + normalize per substrate (react, vue, native, opentui, …) |
 
 ## The map
 
@@ -121,7 +121,7 @@ shared/bindings                       substrate-agnostic event + attr vocabulary
 shared/utils                          cross-target, cross-component helpers
 +-- (composeHandlers, positioning, memo, mergeProps)
 
-<target>                              one substrate (react, native, opentui, …)
+<target>                              one substrate (react, vue, native, opentui, …)
 |                                     runtime, hooks, and props translator
 +-- use-machine                       lifecycle bridge (build + start/stop + useSyncExternalStore)
 +-- use-selector                      fine-grained leaf subscription (O(readers))
@@ -136,7 +136,7 @@ Three package groups, three jobs:
   event + attr vocabulary; `shared/utils` owns agnostic helpers (positioning,
   prop merging, memoization).
 - **`<target>/`** — _the substrate side_. One folder per renderer
-  (`react`, `native`, `opentui`). Owns its runtime bridge and its props translator.
+  (`react`, `vue`, `native`, `opentui`). Owns its runtime bridge and its props translator.
 
 ## The machine parts
 
@@ -174,11 +174,11 @@ whether it needs props/platform or not:
 
 ## Vocabulary
 
-| Term         | What it is                                                                                                                |
-| ------------ | ------------------------------------------------------------------------------------------------------------------------- |
-| **host**     | The agnostic core — `packages/core/*`. Declares what behavior is.                                                         |
-| **target**   | A substrate-specific bridge package and its render environment — `packages/<target>/*` (`react`, `native`, `opentui`, …). |
-| **machine**  | A state-graph config consumed by `machine()`; returns a startable service.                                                |
-| **connect**  | A function returning the logical surface a view spreads onto elements.                                                    |
-| **bindings** | The substrate-agnostic event + attr vocabulary — lives in `shared/bindings`, consumed by every target's normalize.        |
-| **compose**  | Run several machines as one unit (orthogonal regions): bundled `start`/`stop` + `sync` + `combine`.                       |
+| Term         | What it is                                                                                                                       |
+| ------------ | -------------------------------------------------------------------------------------------------------------------------------- |
+| **host**     | The agnostic core — `packages/core/*`. Declares what behavior is.                                                                |
+| **target**   | A substrate-specific bridge package and its render environment — `packages/<target>/*` (`react`, `vue`, `native`, `opentui`, …). |
+| **machine**  | A state-graph config consumed by `machine()`; returns a startable service.                                                       |
+| **connect**  | A function returning the logical surface a view spreads onto elements.                                                           |
+| **bindings** | The substrate-agnostic event + attr vocabulary — lives in `shared/bindings`, consumed by every target's normalize.               |
+| **compose**  | Run several machines as one unit (orthogonal regions): bundled `start`/`stop` + `sync` + `combine`.                              |
