+++ packages/vue/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "@dunky.dev/state-machine-vue",
-  "version": "0.2.0",
+  "name": "@dunky.dev/vue-state-machine",
+  "version": "0.3.2",
   "license": "MIT",
   "repository": {
     "type": "git",
@@ -8,7 +8,8 @@
     "directory": "packages/vue"
   },
   "files": [
-    "dist"
+    "dist",
+    "src"
   ],
   "type": "module",
   "sideEffects": false,
