+++ packages/vue/src/index.ts
@@ -1,4 +1,4 @@
-export { useMachine, type ComponentEffect, type ComponentEffects } from './use-machine'
+export { useMachine, type ComponentEffect } from './use-machine'
 export { useSelector } from './use-selector'
 export { normalize, type Bindings } from './normalize'
 export { mergeProps } from './merge-props'
