+++ sandbox/vue/src/main.ts
@@ -1,5 +1,5 @@
 import { createApp } from 'vue'
-import App from './App.vue'
+import App from './app.vue'
 
 const root = document.getElementById('app')
 if (!root) throw new Error('missing #app')
