+++ sandbox/vue/vite.config.ts
@@ -10,7 +10,7 @@ export default defineConfig({
   resolve: {
     alias: {
       '@dunky.dev/state-machine': resolve(__dirname, '../../packages/core/src'),
-      '@dunky.dev/state-machine-vue': resolve(__dirname, '../../packages/vue/src'),
+      '@dunky.dev/vue-state-machine': resolve(__dirname, '../../packages/vue/src'),
       '@dunky.dev/state-machine-utils': resolve(__dirname, '../../packages/shared/utils/src'),
       '@dunky.dev/state-machine-bindings': resolve(__dirname, '../../packages/shared/bindings/src'),
       '@sandbox/cmdk-core': resolve(__dirname, '../shared/src'),
